Se modifico como se muestra la vida de las construcciones para que no se vean los decimales

+++ src/fiuba/algo3/tp2/Lote.java
@@ -39,7 +39,7 @@ public class Lote {
 		String identpost = "";
 		
 		if (hayConstruccion) {
-			identpost = String.valueOf(construccion.obtenerVida());
+			identpost = String.valueOf((int)construccion.obtenerVida());
 		} else {
 			identpost = "-";
 		}
